+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Stopped a console stream that ended from leaving an online node reported as offline until the browser was reloaded.
+- Renamed the Nodes toolbar "Import" button to "Import server".
+
 ## 1.7.2 - 2026-08-03
 
 - Showed the console command caret.